Module aws_sdk_docdb::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AddSourceIdentifierToSubscription

Fluent builder constructing a request to AddSourceIdentifierToSubscription.

AddTagsToResource

Fluent builder constructing a request to AddTagsToResource.

ApplyPendingMaintenanceAction

Fluent builder constructing a request to ApplyPendingMaintenanceAction.

CopyDBClusterParameterGroup

Fluent builder constructing a request to CopyDBClusterParameterGroup.

CopyDBClusterSnapshot

Fluent builder constructing a request to CopyDBClusterSnapshot.

CreateDBCluster

Fluent builder constructing a request to CreateDBCluster.

CreateDBClusterParameterGroup

Fluent builder constructing a request to CreateDBClusterParameterGroup.

CreateDBClusterSnapshot

Fluent builder constructing a request to CreateDBClusterSnapshot.

CreateDBInstance

Fluent builder constructing a request to CreateDBInstance.

CreateDBSubnetGroup

Fluent builder constructing a request to CreateDBSubnetGroup.

CreateEventSubscription

Fluent builder constructing a request to CreateEventSubscription.

CreateGlobalCluster

Fluent builder constructing a request to CreateGlobalCluster.

DeleteDBCluster

Fluent builder constructing a request to DeleteDBCluster.

DeleteDBClusterParameterGroup

Fluent builder constructing a request to DeleteDBClusterParameterGroup.

DeleteDBClusterSnapshot

Fluent builder constructing a request to DeleteDBClusterSnapshot.

DeleteDBInstance

Fluent builder constructing a request to DeleteDBInstance.

DeleteDBSubnetGroup

Fluent builder constructing a request to DeleteDBSubnetGroup.

DeleteEventSubscription

Fluent builder constructing a request to DeleteEventSubscription.

DeleteGlobalCluster

Fluent builder constructing a request to DeleteGlobalCluster.

DescribeCertificates

Fluent builder constructing a request to DescribeCertificates.

DescribeDBClusterParameterGroups

Fluent builder constructing a request to DescribeDBClusterParameterGroups.

DescribeDBClusterParameters

Fluent builder constructing a request to DescribeDBClusterParameters.

DescribeDBClusterSnapshotAttributes

Fluent builder constructing a request to DescribeDBClusterSnapshotAttributes.

DescribeDBClusterSnapshots

Fluent builder constructing a request to DescribeDBClusterSnapshots.

DescribeDBClusters

Fluent builder constructing a request to DescribeDBClusters.

DescribeDBEngineVersions

Fluent builder constructing a request to DescribeDBEngineVersions.

DescribeDBInstances

Fluent builder constructing a request to DescribeDBInstances.

DescribeDBSubnetGroups

Fluent builder constructing a request to DescribeDBSubnetGroups.

DescribeEngineDefaultClusterParameters

Fluent builder constructing a request to DescribeEngineDefaultClusterParameters.

DescribeEventCategories

Fluent builder constructing a request to DescribeEventCategories.

DescribeEventSubscriptions

Fluent builder constructing a request to DescribeEventSubscriptions.

DescribeEvents

Fluent builder constructing a request to DescribeEvents.

DescribeGlobalClusters

Fluent builder constructing a request to DescribeGlobalClusters.

DescribeOrderableDBInstanceOptions

Fluent builder constructing a request to DescribeOrderableDBInstanceOptions.

DescribePendingMaintenanceActions

Fluent builder constructing a request to DescribePendingMaintenanceActions.

FailoverDBCluster

Fluent builder constructing a request to FailoverDBCluster.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

ModifyDBCluster

Fluent builder constructing a request to ModifyDBCluster.

ModifyDBClusterParameterGroup

Fluent builder constructing a request to ModifyDBClusterParameterGroup.

ModifyDBClusterSnapshotAttribute

Fluent builder constructing a request to ModifyDBClusterSnapshotAttribute.

ModifyDBInstance

Fluent builder constructing a request to ModifyDBInstance.

ModifyDBSubnetGroup

Fluent builder constructing a request to ModifyDBSubnetGroup.

ModifyEventSubscription

Fluent builder constructing a request to ModifyEventSubscription.

ModifyGlobalCluster

Fluent builder constructing a request to ModifyGlobalCluster.

RebootDBInstance

Fluent builder constructing a request to RebootDBInstance.

RemoveFromGlobalCluster

Fluent builder constructing a request to RemoveFromGlobalCluster.

RemoveSourceIdentifierFromSubscription

Fluent builder constructing a request to RemoveSourceIdentifierFromSubscription.

RemoveTagsFromResource

Fluent builder constructing a request to RemoveTagsFromResource.

ResetDBClusterParameterGroup

Fluent builder constructing a request to ResetDBClusterParameterGroup.

RestoreDBClusterFromSnapshot

Fluent builder constructing a request to RestoreDBClusterFromSnapshot.

RestoreDBClusterToPointInTime

Fluent builder constructing a request to RestoreDBClusterToPointInTime.

StartDBCluster

Fluent builder constructing a request to StartDBCluster.

StopDBCluster

Fluent builder constructing a request to StopDBCluster.