Struct aws_sdk_docdb::client::fluent_builders::RestoreDBClusterFromSnapshot

pub struct RestoreDBClusterFromSnapshot<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> RestoreDBClusterFromSnapshot<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<RestoreDbClusterFromSnapshotOutput, SdkError<RestoreDBClusterFromSnapshotError>> where
R::Policy: SmithyRetryPolicy<RestoreDbClusterFromSnapshotInputOperationOutputAlias, RestoreDbClusterFromSnapshotOutput, RestoreDBClusterFromSnapshotError, RestoreDbClusterFromSnapshotInputOperationRetryAlias>,

pub fn availability_zones(self, inp: impl Into<String>) -> Self

Appends an item to AvailabilityZones.

To override the contents of this collection use set_availability_zones.

Provides the list of Amazon EC2 Availability Zones that instances in the restored DB cluster can be created in.

pub fn set_availability_zones(self, input: Option<Vec<String>>) -> Self

pub fn db_cluster_identifier(self, inp: impl Into<String>) -> Self

The name of the cluster to create from the snapshot or cluster snapshot. This parameter isn't case sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-snapshot-id

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn snapshot_identifier(self, inp: impl Into<String>) -> Self

The identifier for the snapshot or cluster snapshot to restore from.

You can use either the name or the Amazon Resource Name (ARN) to specify a cluster snapshot. However, you can use only the ARN to specify a snapshot.

Constraints:

Must match the identifier of an existing snapshot.

pub fn set_snapshot_identifier(self, input: Option<String>) -> Self

pub fn engine(self, inp: impl Into<String>) -> Self

The database engine to use for the new cluster.

Default: The same as source.

Constraint: Must be compatible with the engine of the source.

pub fn set_engine(self, input: Option<String>) -> Self

pub fn engine_version(self, inp: impl Into<String>) -> Self

The version of the database engine to use for the new cluster.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn port(self, inp: i32) -> Self

The port number on which the new cluster accepts connections.

Constraints: Must be a value from 1150 to 65535.

Default: The same port as the original cluster.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn db_subnet_group_name(self, inp: impl Into<String>) -> Self

The name of the subnet group to use for the new cluster.

Constraints: If provided, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

pub fn set_db_subnet_group_name(self, input: Option<String>) -> Self

pub fn vpc_security_group_ids(self, inp: impl Into<String>) -> Self

Appends an item to VpcSecurityGroupIds.

To override the contents of this collection use set_vpc_security_group_ids.

A list of virtual private cloud (VPC) security groups that the new cluster will belong to.

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

The tags to be assigned to the restored cluster.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn kms_key_id(self, inp: impl Into<String>) -> Self

The KMS key identifier to use when restoring an encrypted cluster from a DB snapshot or cluster snapshot.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are restoring a cluster with the same account that owns the KMS encryption key used to encrypt the new cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

If you do not specify a value for the KmsKeyId parameter, then the following occurs:

If the snapshot or cluster snapshot in SnapshotIdentifier is encrypted, then the restored cluster is encrypted using the KMS key that was used to encrypt the snapshot or the cluster snapshot.

If the snapshot or the cluster snapshot in SnapshotIdentifier is not encrypted, then the restored DB cluster is not encrypted.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn enable_cloudwatch_logs_exports(self, inp: impl Into<String>) -> Self

Appends an item to EnableCloudwatchLogsExports.

To override the contents of this collection use set_enable_cloudwatch_logs_exports.

A list of log types that must be enabled for exporting to Amazon CloudWatch Logs.

pub fn set_enable_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

pub fn deletion_protection(self, inp: bool) -> Self

Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self