Struct aws_sdk_docdb::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonRDSv19.

This client allows ergonomic access to a AmazonRDSv19-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn add_source_identifier_to_subscription(
&self
) -> AddSourceIdentifierToSubscription<C, M, R>

Constructs a fluent builder for the AddSourceIdentifierToSubscription operation.

See AddSourceIdentifierToSubscription for more information about the operation and its arguments.

pub fn add_tags_to_resource(&self) -> AddTagsToResource<C, M, R>

Constructs a fluent builder for the AddTagsToResource operation.

See AddTagsToResource for more information about the operation and its arguments.

pub fn apply_pending_maintenance_action(
&self
) -> ApplyPendingMaintenanceAction<C, M, R>

Constructs a fluent builder for the ApplyPendingMaintenanceAction operation.

See ApplyPendingMaintenanceAction for more information about the operation and its arguments.

pub fn copy_db_cluster_parameter_group(
&self
) -> CopyDBClusterParameterGroup<C, M, R>

Constructs a fluent builder for the CopyDBClusterParameterGroup operation.

See CopyDBClusterParameterGroup for more information about the operation and its arguments.

pub fn copy_db_cluster_snapshot(&self) -> CopyDBClusterSnapshot<C, M, R>

Constructs a fluent builder for the CopyDBClusterSnapshot operation.

See CopyDBClusterSnapshot for more information about the operation and its arguments.

pub fn create_db_cluster(&self) -> CreateDBCluster<C, M, R>

Constructs a fluent builder for the CreateDBCluster operation.

See CreateDBCluster for more information about the operation and its arguments.

pub fn create_db_cluster_parameter_group(
&self
) -> CreateDBClusterParameterGroup<C, M, R>

Constructs a fluent builder for the CreateDBClusterParameterGroup operation.

See CreateDBClusterParameterGroup for more information about the operation and its arguments.

pub fn create_db_cluster_snapshot(&self) -> CreateDBClusterSnapshot<C, M, R>

Constructs a fluent builder for the CreateDBClusterSnapshot operation.

See CreateDBClusterSnapshot for more information about the operation and its arguments.

pub fn create_db_instance(&self) -> CreateDBInstance<C, M, R>

Constructs a fluent builder for the CreateDBInstance operation.

See CreateDBInstance for more information about the operation and its arguments.

pub fn create_db_subnet_group(&self) -> CreateDBSubnetGroup<C, M, R>

Constructs a fluent builder for the CreateDBSubnetGroup operation.

See CreateDBSubnetGroup for more information about the operation and its arguments.

pub fn create_event_subscription(&self) -> CreateEventSubscription<C, M, R>

Constructs a fluent builder for the CreateEventSubscription operation.

See CreateEventSubscription for more information about the operation and its arguments.

pub fn create_global_cluster(&self) -> CreateGlobalCluster<C, M, R>

Constructs a fluent builder for the CreateGlobalCluster operation.

See CreateGlobalCluster for more information about the operation and its arguments.

pub fn delete_db_cluster(&self) -> DeleteDBCluster<C, M, R>

Constructs a fluent builder for the DeleteDBCluster operation.

See DeleteDBCluster for more information about the operation and its arguments.

pub fn delete_db_cluster_parameter_group(
&self
) -> DeleteDBClusterParameterGroup<C, M, R>

Constructs a fluent builder for the DeleteDBClusterParameterGroup operation.

See DeleteDBClusterParameterGroup for more information about the operation and its arguments.

pub fn delete_db_cluster_snapshot(&self) -> DeleteDBClusterSnapshot<C, M, R>

Constructs a fluent builder for the DeleteDBClusterSnapshot operation.

See DeleteDBClusterSnapshot for more information about the operation and its arguments.

pub fn delete_db_instance(&self) -> DeleteDBInstance<C, M, R>

Constructs a fluent builder for the DeleteDBInstance operation.

See DeleteDBInstance for more information about the operation and its arguments.

pub fn delete_db_subnet_group(&self) -> DeleteDBSubnetGroup<C, M, R>

Constructs a fluent builder for the DeleteDBSubnetGroup operation.

See DeleteDBSubnetGroup for more information about the operation and its arguments.

pub fn delete_event_subscription(&self) -> DeleteEventSubscription<C, M, R>

Constructs a fluent builder for the DeleteEventSubscription operation.

See DeleteEventSubscription for more information about the operation and its arguments.

pub fn delete_global_cluster(&self) -> DeleteGlobalCluster<C, M, R>

Constructs a fluent builder for the DeleteGlobalCluster operation.

See DeleteGlobalCluster for more information about the operation and its arguments.

pub fn describe_certificates(&self) -> DescribeCertificates<C, M, R>

Constructs a fluent builder for the DescribeCertificates operation.

See DescribeCertificates for more information about the operation and its arguments.

pub fn describe_db_cluster_parameter_groups(
&self
) -> DescribeDBClusterParameterGroups<C, M, R>

Constructs a fluent builder for the DescribeDBClusterParameterGroups operation.

See DescribeDBClusterParameterGroups for more information about the operation and its arguments.

pub fn describe_db_cluster_parameters(
&self
) -> DescribeDBClusterParameters<C, M, R>

Constructs a fluent builder for the DescribeDBClusterParameters operation.

See DescribeDBClusterParameters for more information about the operation and its arguments.

pub fn describe_db_clusters(&self) -> DescribeDBClusters<C, M, R>

Constructs a fluent builder for the DescribeDBClusters operation.

See DescribeDBClusters for more information about the operation and its arguments.

pub fn describe_db_cluster_snapshot_attributes(
&self
) -> DescribeDBClusterSnapshotAttributes<C, M, R>

Constructs a fluent builder for the DescribeDBClusterSnapshotAttributes operation.

See DescribeDBClusterSnapshotAttributes for more information about the operation and its arguments.

pub fn describe_db_cluster_snapshots(
&self
) -> DescribeDBClusterSnapshots<C, M, R>

Constructs a fluent builder for the DescribeDBClusterSnapshots operation.

See DescribeDBClusterSnapshots for more information about the operation and its arguments.

pub fn describe_db_engine_versions(&self) -> DescribeDBEngineVersions<C, M, R>

Constructs a fluent builder for the DescribeDBEngineVersions operation.

See DescribeDBEngineVersions for more information about the operation and its arguments.

pub fn describe_db_instances(&self) -> DescribeDBInstances<C, M, R>

Constructs a fluent builder for the DescribeDBInstances operation.

See DescribeDBInstances for more information about the operation and its arguments.

pub fn describe_db_subnet_groups(&self) -> DescribeDBSubnetGroups<C, M, R>

Constructs a fluent builder for the DescribeDBSubnetGroups operation.

See DescribeDBSubnetGroups for more information about the operation and its arguments.

pub fn describe_engine_default_cluster_parameters(
&self
) -> DescribeEngineDefaultClusterParameters<C, M, R>

Constructs a fluent builder for the DescribeEngineDefaultClusterParameters operation.

See DescribeEngineDefaultClusterParameters for more information about the operation and its arguments.

pub fn describe_event_categories(&self) -> DescribeEventCategories<C, M, R>

Constructs a fluent builder for the DescribeEventCategories operation.

See DescribeEventCategories for more information about the operation and its arguments.

pub fn describe_events(&self) -> DescribeEvents<C, M, R>

Constructs a fluent builder for the DescribeEvents operation.

See DescribeEvents for more information about the operation and its arguments.

pub fn describe_event_subscriptions(
&self
) -> DescribeEventSubscriptions<C, M, R>

Constructs a fluent builder for the DescribeEventSubscriptions operation.

See DescribeEventSubscriptions for more information about the operation and its arguments.

pub fn describe_global_clusters(&self) -> DescribeGlobalClusters<C, M, R>

Constructs a fluent builder for the DescribeGlobalClusters operation.

See DescribeGlobalClusters for more information about the operation and its arguments.

pub fn describe_orderable_db_instance_options(
&self
) -> DescribeOrderableDBInstanceOptions<C, M, R>

Constructs a fluent builder for the DescribeOrderableDBInstanceOptions operation.

See DescribeOrderableDBInstanceOptions for more information about the operation and its arguments.

pub fn describe_pending_maintenance_actions(
&self
) -> DescribePendingMaintenanceActions<C, M, R>

Constructs a fluent builder for the DescribePendingMaintenanceActions operation.

See DescribePendingMaintenanceActions for more information about the operation and its arguments.

pub fn failover_db_cluster(&self) -> FailoverDBCluster<C, M, R>

Constructs a fluent builder for the FailoverDBCluster operation.

See FailoverDBCluster for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn modify_db_cluster(&self) -> ModifyDBCluster<C, M, R>

Constructs a fluent builder for the ModifyDBCluster operation.

See ModifyDBCluster for more information about the operation and its arguments.

pub fn modify_db_cluster_parameter_group(
&self
) -> ModifyDBClusterParameterGroup<C, M, R>

Constructs a fluent builder for the ModifyDBClusterParameterGroup operation.

See ModifyDBClusterParameterGroup for more information about the operation and its arguments.

pub fn modify_db_cluster_snapshot_attribute(
&self
) -> ModifyDBClusterSnapshotAttribute<C, M, R>

Constructs a fluent builder for the ModifyDBClusterSnapshotAttribute operation.

See ModifyDBClusterSnapshotAttribute for more information about the operation and its arguments.

pub fn modify_db_instance(&self) -> ModifyDBInstance<C, M, R>

Constructs a fluent builder for the ModifyDBInstance operation.

See ModifyDBInstance for more information about the operation and its arguments.

pub fn modify_db_subnet_group(&self) -> ModifyDBSubnetGroup<C, M, R>

Constructs a fluent builder for the ModifyDBSubnetGroup operation.

See ModifyDBSubnetGroup for more information about the operation and its arguments.

pub fn modify_event_subscription(&self) -> ModifyEventSubscription<C, M, R>

Constructs a fluent builder for the ModifyEventSubscription operation.

See ModifyEventSubscription for more information about the operation and its arguments.

pub fn modify_global_cluster(&self) -> ModifyGlobalCluster<C, M, R>

Constructs a fluent builder for the ModifyGlobalCluster operation.

See ModifyGlobalCluster for more information about the operation and its arguments.

pub fn reboot_db_instance(&self) -> RebootDBInstance<C, M, R>

Constructs a fluent builder for the RebootDBInstance operation.

See RebootDBInstance for more information about the operation and its arguments.

pub fn remove_from_global_cluster(&self) -> RemoveFromGlobalCluster<C, M, R>

Constructs a fluent builder for the RemoveFromGlobalCluster operation.

See RemoveFromGlobalCluster for more information about the operation and its arguments.

pub fn remove_source_identifier_from_subscription(
&self
) -> RemoveSourceIdentifierFromSubscription<C, M, R>

Constructs a fluent builder for the RemoveSourceIdentifierFromSubscription operation.

See RemoveSourceIdentifierFromSubscription for more information about the operation and its arguments.

pub fn remove_tags_from_resource(&self) -> RemoveTagsFromResource<C, M, R>

Constructs a fluent builder for the RemoveTagsFromResource operation.

See RemoveTagsFromResource for more information about the operation and its arguments.

pub fn reset_db_cluster_parameter_group(
&self
) -> ResetDBClusterParameterGroup<C, M, R>

Constructs a fluent builder for the ResetDBClusterParameterGroup operation.

See ResetDBClusterParameterGroup for more information about the operation and its arguments.

pub fn restore_db_cluster_from_snapshot(
&self
) -> RestoreDBClusterFromSnapshot<C, M, R>

Constructs a fluent builder for the RestoreDBClusterFromSnapshot operation.

See RestoreDBClusterFromSnapshot for more information about the operation and its arguments.

pub fn restore_db_cluster_to_point_in_time(
&self
) -> RestoreDBClusterToPointInTime<C, M, R>

Constructs a fluent builder for the RestoreDBClusterToPointInTime operation.

See RestoreDBClusterToPointInTime for more information about the operation and its arguments.

pub fn start_db_cluster(&self) -> StartDBCluster<C, M, R>

Constructs a fluent builder for the StartDBCluster operation.

See StartDBCluster for more information about the operation and its arguments.

pub fn stop_db_cluster(&self) -> StopDBCluster<C, M, R>

Constructs a fluent builder for the StopDBCluster operation.

See StopDBCluster for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.