Module aws_sdk_docdb::error

Modules

authorization_not_found_fault

See AuthorizationNotFoundFault

certificate_not_found_fault

See CertificateNotFoundFault

db_cluster_already_exists_fault

See DbClusterAlreadyExistsFault

db_cluster_not_found_fault

See DbClusterNotFoundFault

db_cluster_parameter_group_not_found_fault

See DbClusterParameterGroupNotFoundFault

db_cluster_quota_exceeded_fault

See DbClusterQuotaExceededFault

db_cluster_snapshot_already_exists_fault

See DbClusterSnapshotAlreadyExistsFault

db_cluster_snapshot_not_found_fault

See DbClusterSnapshotNotFoundFault

db_instance_already_exists_fault

See DbInstanceAlreadyExistsFault

db_instance_not_found_fault

See DbInstanceNotFoundFault

db_parameter_group_already_exists_fault

See DbParameterGroupAlreadyExistsFault

db_parameter_group_not_found_fault

See DbParameterGroupNotFoundFault

db_parameter_group_quota_exceeded_fault

See DbParameterGroupQuotaExceededFault

db_security_group_not_found_fault

See DbSecurityGroupNotFoundFault

db_snapshot_already_exists_fault

See DbSnapshotAlreadyExistsFault

db_snapshot_not_found_fault

See DbSnapshotNotFoundFault

db_subnet_group_already_exists_fault

See DbSubnetGroupAlreadyExistsFault

db_subnet_group_does_not_cover_enough_a_zs

See DbSubnetGroupDoesNotCoverEnoughAZs

db_subnet_group_not_found_fault

See DbSubnetGroupNotFoundFault

db_subnet_group_quota_exceeded_fault

See DbSubnetGroupQuotaExceededFault

db_subnet_quota_exceeded_fault

See DbSubnetQuotaExceededFault

db_upgrade_dependency_failure_fault

See DbUpgradeDependencyFailureFault

event_subscription_quota_exceeded_fault

See EventSubscriptionQuotaExceededFault

global_cluster_already_exists_fault

See GlobalClusterAlreadyExistsFault

global_cluster_not_found_fault

See GlobalClusterNotFoundFault

global_cluster_quota_exceeded_fault

See GlobalClusterQuotaExceededFault

instance_quota_exceeded_fault

See InstanceQuotaExceededFault

insufficient_db_cluster_capacity_fault

See InsufficientDbClusterCapacityFault

insufficient_db_instance_capacity_fault

See InsufficientDbInstanceCapacityFault

insufficient_storage_cluster_capacity_fault

See InsufficientStorageClusterCapacityFault

invalid_db_cluster_snapshot_state_fault

See InvalidDbClusterSnapshotStateFault

invalid_db_cluster_state_fault

See InvalidDbClusterStateFault

invalid_db_instance_state_fault

See InvalidDbInstanceStateFault

invalid_db_parameter_group_state_fault

See InvalidDbParameterGroupStateFault

invalid_db_security_group_state_fault

See InvalidDbSecurityGroupStateFault

invalid_db_snapshot_state_fault

See InvalidDbSnapshotStateFault

invalid_db_subnet_group_state_fault

See InvalidDbSubnetGroupStateFault

invalid_db_subnet_state_fault

See InvalidDbSubnetStateFault

invalid_event_subscription_state_fault

See InvalidEventSubscriptionStateFault

invalid_global_cluster_state_fault

See InvalidGlobalClusterStateFault

invalid_restore_fault

See InvalidRestoreFault

invalid_subnet

See InvalidSubnet

invalid_vpc_network_state_fault

See InvalidVpcNetworkStateFault

kms_key_not_accessible_fault

See KmsKeyNotAccessibleFault

resource_not_found_fault

See ResourceNotFoundFault

shared_snapshot_quota_exceeded_fault

See SharedSnapshotQuotaExceededFault

snapshot_quota_exceeded_fault

See SnapshotQuotaExceededFault

sns_invalid_topic_fault

See SnsInvalidTopicFault

sns_no_authorization_fault

See SnsNoAuthorizationFault

sns_topic_arn_not_found_fault

See SnsTopicArnNotFoundFault

source_not_found_fault

See SourceNotFoundFault

storage_quota_exceeded_fault

See StorageQuotaExceededFault

storage_type_not_supported_fault

See StorageTypeNotSupportedFault

subnet_already_in_use

See SubnetAlreadyInUse

subscription_already_exist_fault

See SubscriptionAlreadyExistFault

subscription_category_not_found_fault

See SubscriptionCategoryNotFoundFault

subscription_not_found_fault

See SubscriptionNotFoundFault

Structs

AddSourceIdentifierToSubscriptionError

AddTagsToResourceError

ApplyPendingMaintenanceActionError

AuthorizationNotFoundFault

The specified CIDR IP or Amazon EC2 security group isn't authorized for the specified security group.

CertificateNotFoundFault

CopyDBClusterParameterGroupError

CopyDBClusterSnapshotError

CreateDBClusterError

CreateDBClusterParameterGroupError

CreateDBClusterSnapshotError

CreateDBInstanceError

CreateDBSubnetGroupError

CreateEventSubscriptionError

CreateGlobalClusterError

DbClusterAlreadyExistsFault

You already have a cluster with the given identifier.

DbClusterNotFoundFault

DbClusterParameterGroupNotFoundFault

DbClusterQuotaExceededFault

The cluster can't be created because you have reached the maximum allowed quota of clusters.

DbClusterSnapshotAlreadyExistsFault

You already have a cluster snapshot with the given identifier.

DbClusterSnapshotNotFoundFault

DbInstanceAlreadyExistsFault

You already have a instance with the given identifier.

DbInstanceNotFoundFault

DbParameterGroupAlreadyExistsFault

A parameter group with the same name already exists.

DbParameterGroupNotFoundFault

DbParameterGroupQuotaExceededFault

This request would cause you to exceed the allowed number of parameter groups.

DbSecurityGroupNotFoundFault

DbSnapshotAlreadyExistsFault

DbSnapshotNotFoundFault

DbSubnetGroupAlreadyExistsFault

DbSubnetGroupDoesNotCoverEnoughAZs

Subnets in the subnet group should cover at least two Availability Zones unless there is only one Availability Zone.

DbSubnetGroupNotFoundFault

DbSubnetGroupQuotaExceededFault

The request would cause you to exceed the allowed number of subnet groups.

DbSubnetQuotaExceededFault

The request would cause you to exceed the allowed number of subnets in a subnet group.

DbUpgradeDependencyFailureFault

The upgrade failed because a resource that the depends on can't be

DeleteDBClusterError

DeleteDBClusterParameterGroupError

DeleteDBClusterSnapshotError

DeleteDBInstanceError

DeleteDBSubnetGroupError

DeleteEventSubscriptionError

DeleteGlobalClusterError

DescribeCertificatesError

DescribeDBClusterParameterGroupsError

DescribeDBClusterParametersError

DescribeDBClusterSnapshotAttributesError

DescribeDBClusterSnapshotsError

DescribeDBClustersError

DescribeDBEngineVersionsError

DescribeDBInstancesError

DescribeDBSubnetGroupsError

DescribeEngineDefaultClusterParametersError

DescribeEventCategoriesError

DescribeEventSubscriptionsError

DescribeEventsError

DescribeGlobalClustersError

DescribeOrderableDBInstanceOptionsError

DescribePendingMaintenanceActionsError

EventSubscriptionQuotaExceededFault

You have reached the maximum number of event subscriptions.

FailoverDBClusterError

GlobalClusterAlreadyExistsFault

The GlobalClusterIdentifier already exists. Choose a new global cluster identifier (unique name) to create a new global cluster.

GlobalClusterNotFoundFault

The GlobalClusterIdentifier doesn't refer to an existing global cluster.

GlobalClusterQuotaExceededFault

The number of global clusters for this account is already at the maximum allowed.

InstanceQuotaExceededFault

The request would cause you to exceed the allowed number of instances.

InsufficientDbClusterCapacityFault

The cluster doesn't have enough capacity for the current operation.

InsufficientDbInstanceCapacityFault

The specified instance class isn't available in the specified Availability Zone.

InsufficientStorageClusterCapacityFault

There is not enough storage available for the current action. You might be able to resolve this error by updating your subnet group to use different Availability Zones that have more storage available.

InvalidDbClusterSnapshotStateFault

The provided value isn't a valid cluster snapshot state.

InvalidDbClusterStateFault

The cluster isn't in a valid state.

InvalidDbInstanceStateFault

The specified instance isn't in the available state.

InvalidDbParameterGroupStateFault

The parameter group is in use, or it is in a state that is not valid. If you are trying to delete the parameter group, you can't delete it when the parameter group is in this state.

InvalidDbSecurityGroupStateFault

The state of the security group doesn't allow deletion.

InvalidDbSnapshotStateFault

The state of the snapshot doesn't allow deletion.

InvalidDbSubnetGroupStateFault

The subnet group can't be deleted because it's in use.

InvalidDbSubnetStateFault

The subnet isn't in the available state.

InvalidEventSubscriptionStateFault

Someone else might be modifying a subscription. Wait a few seconds, and try

InvalidGlobalClusterStateFault

The requested operation can't be performed while the cluster is in this state.

InvalidRestoreFault

You cannot restore from a virtual private cloud (VPC) backup to a non-VPC DB

InvalidSubnet

The requested subnet is not valid, or multiple subnets were requested that are not all

InvalidVpcNetworkStateFault

The subnet group doesn't cover all Availability Zones after it is created

KmsKeyNotAccessibleFault

An error occurred when accessing an KMS key.

ListTagsForResourceError

ModifyDBClusterError

ModifyDBClusterParameterGroupError

ModifyDBClusterSnapshotAttributeError

ModifyDBInstanceError

ModifyDBSubnetGroupError

ModifyEventSubscriptionError

ModifyGlobalClusterError

RebootDBInstanceError

RemoveFromGlobalClusterError

RemoveSourceIdentifierFromSubscriptionError

RemoveTagsFromResourceError

ResetDBClusterParameterGroupError

ResourceNotFoundFault

The specified resource ID was not found.

RestoreDBClusterFromSnapshotError

RestoreDBClusterToPointInTimeError

SharedSnapshotQuotaExceededFault

You have exceeded the maximum number of accounts that you can share a manual DB

SnapshotQuotaExceededFault

The request would cause you to exceed the allowed number of snapshots.

SnsInvalidTopicFault

Amazon SNS has responded that there is a problem with the specified topic.

SnsNoAuthorizationFault

You do not have permission to publish to the SNS topic Amazon Resource Name (ARN).

SnsTopicArnNotFoundFault

The SNS topic Amazon Resource Name (ARN) does not exist.

SourceNotFoundFault

The requested source could not be found.

StartDBClusterError

StopDBClusterError

StorageQuotaExceededFault

The request would cause you to exceed the allowed amount of storage available across

StorageTypeNotSupportedFault

Storage of the specified StorageType can't be associated with the DB

SubnetAlreadyInUse

The subnet is already in use in the Availability Zone.

SubscriptionAlreadyExistFault

The provided subscription name already exists.

SubscriptionCategoryNotFoundFault

The provided category does not exist.

SubscriptionNotFoundFault

The subscription name does not exist.

Enums

AddSourceIdentifierToSubscriptionErrorKind

AddTagsToResourceErrorKind

ApplyPendingMaintenanceActionErrorKind

CopyDBClusterParameterGroupErrorKind

CopyDBClusterSnapshotErrorKind

CreateDBClusterErrorKind

CreateDBClusterParameterGroupErrorKind

CreateDBClusterSnapshotErrorKind

CreateDBInstanceErrorKind

CreateDBSubnetGroupErrorKind

CreateEventSubscriptionErrorKind

CreateGlobalClusterErrorKind

DeleteDBClusterErrorKind

DeleteDBClusterParameterGroupErrorKind

DeleteDBClusterSnapshotErrorKind

DeleteDBInstanceErrorKind

DeleteDBSubnetGroupErrorKind

DeleteEventSubscriptionErrorKind

DeleteGlobalClusterErrorKind

DescribeCertificatesErrorKind

DescribeDBClusterParameterGroupsErrorKind

DescribeDBClusterParametersErrorKind

DescribeDBClusterSnapshotAttributesErrorKind

DescribeDBClusterSnapshotsErrorKind

DescribeDBClustersErrorKind

DescribeDBEngineVersionsErrorKind

DescribeDBInstancesErrorKind

DescribeDBSubnetGroupsErrorKind

DescribeEngineDefaultClusterParametersErrorKind

DescribeEventCategoriesErrorKind

DescribeEventSubscriptionsErrorKind

DescribeEventsErrorKind

DescribeGlobalClustersErrorKind

DescribeOrderableDBInstanceOptionsErrorKind

DescribePendingMaintenanceActionsErrorKind

FailoverDBClusterErrorKind

ListTagsForResourceErrorKind

ModifyDBClusterErrorKind

ModifyDBClusterParameterGroupErrorKind

ModifyDBClusterSnapshotAttributeErrorKind

ModifyDBInstanceErrorKind

ModifyDBSubnetGroupErrorKind

ModifyEventSubscriptionErrorKind

ModifyGlobalClusterErrorKind

RebootDBInstanceErrorKind

RemoveFromGlobalClusterErrorKind

RemoveSourceIdentifierFromSubscriptionErrorKind

RemoveTagsFromResourceErrorKind

ResetDBClusterParameterGroupErrorKind

RestoreDBClusterFromSnapshotErrorKind

RestoreDBClusterToPointInTimeErrorKind

StartDBClusterErrorKind

StopDBClusterErrorKind