Module aws_sdk_docdb::model

Modules

availability_zone

See AvailabilityZone

certificate

See Certificate

cloudwatch_logs_export_configuration

See CloudwatchLogsExportConfiguration

db_cluster

See DbCluster

db_cluster_member

See DbClusterMember

db_cluster_parameter_group

See DbClusterParameterGroup

db_cluster_role

See DbClusterRole

db_cluster_snapshot

See DbClusterSnapshot

db_cluster_snapshot_attribute

See DbClusterSnapshotAttribute

db_cluster_snapshot_attributes_result

See DbClusterSnapshotAttributesResult

db_engine_version

See DbEngineVersion

db_instance

See DbInstance

db_instance_status_info

See DbInstanceStatusInfo

db_subnet_group

See DbSubnetGroup

endpoint

See Endpoint

engine_defaults

See EngineDefaults

event

See Event

event_categories_map

See EventCategoriesMap

event_subscription

See EventSubscription

filter

See Filter

global_cluster

See GlobalCluster

global_cluster_member

See GlobalClusterMember

orderable_db_instance_option

See OrderableDbInstanceOption

parameter

See Parameter

pending_cloudwatch_logs_exports

See PendingCloudwatchLogsExports

pending_maintenance_action

See PendingMaintenanceAction

pending_modified_values

See PendingModifiedValues

resource_pending_maintenance_actions

See ResourcePendingMaintenanceActions

subnet

See Subnet

tag

See Tag

upgrade_target

See UpgradeTarget

vpc_security_group_membership

See VpcSecurityGroupMembership

Structs

AvailabilityZone

Information about an Availability Zone.

Certificate

A certificate authority (CA) certificate for an account.

CloudwatchLogsExportConfiguration

The configuration setting for the log types to be enabled for export to Amazon

DbCluster

Detailed information about a cluster.

DbClusterMember

Contains information about an instance that is part of a cluster.

DbClusterParameterGroup

Detailed information about a cluster parameter group.

DbClusterRole

Describes an Identity and Access Management (IAM) role that is associated with a

DbClusterSnapshot

Detailed information about a cluster snapshot.

DbClusterSnapshotAttribute

Contains the name and values of a manual cluster snapshot attribute.

DbClusterSnapshotAttributesResult

Detailed information about the attributes that are associated with a cluster

DbEngineVersion

Detailed information about an engine version.

DbInstance

Detailed information about an instance.

DbInstanceStatusInfo

Provides a list of status information for an instance.

DbSubnetGroup

Detailed information about a subnet group.

Endpoint

Network information for accessing a cluster or instance. Client programs must

EngineDefaults

Contains the result of a successful invocation of the

Event

Detailed information about an event.

EventCategoriesMap

An event source type, accompanied by one or more event category names.

EventSubscription

Detailed information about an event to which you have subscribed.

Filter

A named set of filter values, used to return a more specific list of results. You can

GlobalCluster

A data type representing an Amazon DocumentDB global cluster.

GlobalClusterMember

A data structure with information about any primary and secondary clusters associated with an Amazon DocumentDB global clusters.

OrderableDbInstanceOption

The options that are available for an instance.

Parameter

Detailed information about an individual parameter.

PendingCloudwatchLogsExports

A list of the log types whose configuration is still pending. These log types are in

PendingMaintenanceAction

Provides information about a pending maintenance action for a resource.

PendingModifiedValues

One or more modified settings for an instance. These modified settings have been

ResourcePendingMaintenanceActions

Represents the output of ApplyPendingMaintenanceAction.

Subnet

Detailed information about a subnet.

Tag

Metadata assigned to an Amazon DocumentDB resource consisting of a key-value pair.

UpgradeTarget

The version of the database engine that an instance can be upgraded to.

VpcSecurityGroupMembership

Used as a response element for queries on virtual private cloud (VPC) security group

Enums

ApplyMethod

SourceType