Struct aws_sdk_docdb::model::db_cluster::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DbCluster

Implementations

impl Builder

pub fn availability_zones(self, input: impl Into<String>) -> Self

pub fn set_availability_zones(self, input: Option<Vec<String>>) -> Self

pub fn backup_retention_period(self, input: i32) -> Self

Specifies the number of days for which automatic snapshots are retained.

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

pub fn db_cluster_identifier(self, input: impl Into<String>) -> Self

Contains a user-supplied cluster identifier. This identifier is the unique key that identifies a cluster.

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn db_cluster_parameter_group(self, input: impl Into<String>) -> Self

Specifies the name of the cluster parameter group for the cluster.

pub fn set_db_cluster_parameter_group(self, input: Option<String>) -> Self

pub fn db_subnet_group(self, input: impl Into<String>) -> Self

Specifies information on the subnet group that is associated with the cluster, including the name, description, and subnets in the subnet group.

pub fn set_db_subnet_group(self, input: Option<String>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

Specifies the current state of this cluster.

pub fn set_status(self, input: Option<String>) -> Self

pub fn percent_progress(self, input: impl Into<String>) -> Self

Specifies the progress of the operation as a percentage.

pub fn set_percent_progress(self, input: Option<String>) -> Self

pub fn earliest_restorable_time(self, input: Instant) -> Self

The earliest time to which a database can be restored with point-in-time restore.

pub fn set_earliest_restorable_time(self, input: Option<Instant>) -> Self

pub fn endpoint(self, input: impl Into<String>) -> Self

Specifies the connection endpoint for the primary instance of the cluster.

pub fn set_endpoint(self, input: Option<String>) -> Self

pub fn reader_endpoint(self, input: impl Into<String>) -> Self

The reader endpoint for the cluster. The reader endpoint for a cluster load balances connections across the Amazon DocumentDB replicas that are available in a cluster. As clients request new connections to the reader endpoint, Amazon DocumentDB distributes the connection requests among the Amazon DocumentDB replicas in the cluster. This functionality can help balance your read workload across multiple Amazon DocumentDB replicas in your cluster.

If a failover occurs, and the Amazon DocumentDB replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Amazon DocumentDB replicas in the cluster, you can then reconnect to the reader endpoint.

pub fn set_reader_endpoint(self, input: Option<String>) -> Self

pub fn multi_az(self, input: bool) -> Self

Specifies whether the cluster has instances in multiple Availability Zones.

pub fn set_multi_az(self, input: Option<bool>) -> Self

pub fn engine(self, input: impl Into<String>) -> Self

Provides the name of the database engine to be used for this cluster.

pub fn set_engine(self, input: Option<String>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

Indicates the database engine version.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn latest_restorable_time(self, input: Instant) -> Self

Specifies the latest time to which a database can be restored with point-in-time restore.

pub fn set_latest_restorable_time(self, input: Option<Instant>) -> Self

pub fn port(self, input: i32) -> Self

Specifies the port that the database engine is listening on.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn master_username(self, input: impl Into<String>) -> Self

Contains the master user name for the cluster.

pub fn set_master_username(self, input: Option<String>) -> Self

pub fn preferred_backup_window(self, input: impl Into<String>) -> Self

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn replication_source_identifier(self, input: impl Into<String>) -> Self

Contains the identifier of the source cluster if this cluster is a secondary cluster.

pub fn set_replication_source_identifier(self, input: Option<String>) -> Self

pub fn read_replica_identifiers(self, input: impl Into<String>) -> Self

pub fn set_read_replica_identifiers(self, input: Option<Vec<String>>) -> Self

pub fn db_cluster_members(self, input: impl Into<DbClusterMember>) -> Self

pub fn set_db_cluster_members(self, input: Option<Vec<DbClusterMember>>) -> Self

pub fn vpc_security_groups(
self,
input: impl Into<VpcSecurityGroupMembership>
) -> Self

pub fn set_vpc_security_groups(
self,
input: Option<Vec<VpcSecurityGroupMembership>>
) -> Self

pub fn hosted_zone_id(self, input: impl Into<String>) -> Self

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

pub fn set_hosted_zone_id(self, input: Option<String>) -> Self

pub fn storage_encrypted(self, input: bool) -> Self

Specifies whether the cluster is encrypted.

pub fn set_storage_encrypted(self, input: Option<bool>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

If StorageEncrypted is true, the KMS key identifier for the encrypted cluster.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn db_cluster_resource_id(self, input: impl Into<String>) -> Self

The Region-unique, immutable identifier for the cluster. This identifier is found in CloudTrail log entries whenever the KMS key for the cluster is accessed.

pub fn set_db_cluster_resource_id(self, input: Option<String>) -> Self

pub fn db_cluster_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) for the cluster.

pub fn set_db_cluster_arn(self, input: Option<String>) -> Self

pub fn associated_roles(self, input: impl Into<DbClusterRole>) -> Self

pub fn set_associated_roles(self, input: Option<Vec<DbClusterRole>>) -> Self

pub fn cluster_create_time(self, input: Instant) -> Self

Specifies the time when the cluster was created, in Universal Coordinated Time (UTC).

pub fn set_cluster_create_time(self, input: Option<Instant>) -> Self

pub fn enabled_cloudwatch_logs_exports(self, input: impl Into<String>) -> Self

pub fn set_enabled_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

pub fn deletion_protection(self, input: bool) -> Self

Specifies whether this cluster can be deleted. If DeletionProtection is enabled, the cluster cannot be deleted unless it is modified and DeletionProtection is disabled. DeletionProtection protects clusters from being accidentally deleted.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

pub fn build(self) -> DbCluster

Consumes the builder and constructs a DbCluster