Module aws_sdk_docdb::operation

Structs

AddSourceIdentifierToSubscription

Adds a source identifier to an existing event notification

AddTagsToResource

Adds metadata tags to an Amazon DocumentDB resource. You can use these tags

ApplyPendingMaintenanceAction

Applies a pending maintenance action to a resource (for example,

CopyDBClusterParameterGroup

Copies the specified cluster parameter group.

CopyDBClusterSnapshot

Copies a snapshot of a cluster.

CreateDBCluster

Creates a new Amazon DocumentDB cluster.

CreateDBClusterParameterGroup

Creates a new cluster parameter group.

CreateDBClusterSnapshot

Creates a snapshot of a cluster.

CreateDBInstance

Creates a new instance.

CreateDBSubnetGroup

Creates a new subnet group. subnet groups must contain at least one subnet in at

CreateEventSubscription

Creates an Amazon DocumentDB event notification subscription. This action requires a topic Amazon Resource Name (ARN) created by using the Amazon DocumentDB console, the Amazon SNS console, or the Amazon SNS API. To obtain an ARN with Amazon SNS, you must create a topic in Amazon SNS and subscribe to the topic. The ARN is displayed in the Amazon SNS console.

CreateGlobalCluster

Creates an Amazon DocumentDB global cluster that can span multiple multiple Regions. The global cluster contains one primary cluster with read-write capability, and up-to give read-only secondary clusters. Global clusters uses storage-based fast replication across regions with latencies less than one second, using dedicated infrastructure with no impact to your workload’s performance.

DeleteDBCluster

Deletes a previously provisioned cluster. When you delete a cluster, all automated backups for that cluster are deleted and can't be recovered. Manual DB cluster snapshots of the specified cluster are not deleted.

DeleteDBClusterParameterGroup

Deletes a specified cluster parameter group. The cluster parameter group to be deleted can't be associated with any clusters.

DeleteDBClusterSnapshot

Deletes a cluster snapshot. If the snapshot is being copied, the copy operation is terminated.

DeleteDBInstance

Deletes a previously provisioned instance.

DeleteDBSubnetGroup

Deletes a subnet group.

DeleteEventSubscription

Deletes an Amazon DocumentDB event notification subscription.

DeleteGlobalCluster

Deletes a global cluster. The primary and secondary clusters must already be detached or deleted before attempting to delete a global cluster.

DescribeCertificates

Returns a list of certificate authority (CA) certificates provided by Amazon DocumentDB for this account.

DescribeDBClusterParameterGroups

Returns a list of DBClusterParameterGroup descriptions. If a DBClusterParameterGroupName parameter is specified, the list contains only the description of the specified cluster parameter group.

DescribeDBClusterParameters

Returns the detailed parameter list for a particular cluster parameter

DescribeDBClusterSnapshotAttributes

Returns a list of cluster snapshot attribute names and values for a manual DB

DescribeDBClusterSnapshots

Returns information about cluster snapshots. This API operation supports pagination.

DescribeDBClusters

Returns information about provisioned Amazon DocumentDB clusters. This API

DescribeDBEngineVersions

Returns a list of the available engines.

DescribeDBInstances

Returns information about provisioned Amazon DocumentDB instances. This API supports pagination.

DescribeDBSubnetGroups

Returns a list of DBSubnetGroup descriptions. If a

DescribeEngineDefaultClusterParameters

Returns the default engine and system parameter information for the cluster database

DescribeEventCategories

Displays a list of categories for all event source types, or, if specified, for a

DescribeEventSubscriptions

Lists all the subscription descriptions for a customer account. The description for a subscription includes SubscriptionName, SNSTopicARN, CustomerID, SourceType, SourceID, CreationTime, and Status.

DescribeEvents

Returns events related to instances, security groups, snapshots, and DB parameter groups for the past 14 days. You can obtain events specific to a particular DB instance, security group, snapshot, or parameter group by providing the name as a parameter. By default, the events of the past hour are returned.

DescribeGlobalClusters

Returns information about Amazon DocumentDB global clusters. This API supports pagination.

DescribeOrderableDBInstanceOptions

Returns a list of orderable instance options for the specified engine.

DescribePendingMaintenanceActions

Returns a list of resources (for example, instances) that have at least one pending

FailoverDBCluster

Forces a failover for a cluster.

ListTagsForResource

Lists all tags on an Amazon DocumentDB resource.

ModifyDBCluster

Modifies a setting for an Amazon DocumentDB cluster. You can change one or more database

ModifyDBClusterParameterGroup

Modifies the parameters of a cluster parameter group. To modify more than one

ModifyDBClusterSnapshotAttribute

Adds an attribute and values to, or removes an attribute and values from, a manual cluster snapshot.

ModifyDBInstance

Modifies settings for an instance. You can change one or more database configuration parameters by specifying these parameters and the new values in the request.

ModifyDBSubnetGroup

Modifies an existing subnet group. subnet groups must contain at least one subnet in at least two Availability Zones in the Region.

ModifyEventSubscription

Modifies an existing Amazon DocumentDB event notification subscription.

ModifyGlobalCluster

Modify a setting for an Amazon DocumentDB global cluster. You can change one or more configuration parameters (for example: deletion protection), or the global cluster identifier by specifying these parameters and the new values in the request.

RebootDBInstance

You might need to reboot your instance, usually for maintenance reasons. For

RemoveFromGlobalCluster

Detaches an Amazon DocumentDB secondary cluster from a global cluster. The cluster becomes a standalone cluster with read-write capability instead of being read-only and receiving data from a primary in a different region.

RemoveSourceIdentifierFromSubscription

Removes a source identifier from an existing Amazon DocumentDB event notification

RemoveTagsFromResource

Removes metadata tags from an Amazon DocumentDB resource.

ResetDBClusterParameterGroup

Modifies the parameters of a cluster parameter group to the default value. To

RestoreDBClusterFromSnapshot

Creates a new cluster from a snapshot or cluster snapshot.

RestoreDBClusterToPointInTime

Restores a cluster to an arbitrary point in time. Users can restore to any point in

StartDBCluster

Restarts the stopped cluster that is specified by DBClusterIdentifier.

StopDBCluster

Stops the running cluster that is specified by DBClusterIdentifier. The